Move GoalView default state to separate file
