+++ src/components/GoalView/GoalSelector/tests/GoalSelectorComponent.test.tsx
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom";
 import { act } from "react-dom/test-utils";
 import configureMockStore from "redux-mock-store";
 import GoalSelector from "../";
-import { defaultState } from "../../GoalViewReducers";
+import { defaultState } from "../../TempDefaultState";
 import { Provider } from "react-redux";
 
 const createMockStore = configureMockStore([]);
