+++ src/components/GoalView/tests/GoalView.test.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import configureMockStore from "redux-mock-store";
-import { defaultState } from "../GoalViewReducers";
+import { defaultState } from "../TempDefaultState";
 import { Provider } from "react-redux";
 import { GoalView } from "../GoalView";
